+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## UNRELEASED
+
+- Include CGI/HTTP request rack environment variables in the JSON dump.
+
 ## 7.0.0 (2024-02-29)
 
 - Add the `--profile` CLI option. This option will print the names of the 10